+++ excelerator/__init__.py
@@ -2,8 +2,7 @@
 Library for reading tabular data from excel in a very forgiving manner.
 """
 
-__version__ = "0.3.1"
+__version__ = "0.4"
 
 
-from excelerator.table_reader import TableReader
-from excelerator import exceptions
+from excelerator.main.table_reader import TableReader